planner/core: code refactor around join reorder (#9439)
