+++ planner/core/rule_join_reorder_greedy.go
@@ -19,48 +19,11 @@ import (
 
 	"github.com/pingcap/parser/ast"
 	"github.com/pingcap/tidb/expression"
-	"github.com/pingcap/tidb/sessionctx"
 )
 
-// extractJoinGroup extracts all the join nodes connected with continuous
-// InnerJoins to construct a join group. This join group is further used to
-// construct a new join order based on a greedy algorithm.
-//
-// For example: "InnerJoin(InnerJoin(a, b), LeftJoin(c, d))"
-// results in a join group {a, b, LeftJoin(c, d)}.
-func extractJoinGroup(p LogicalPlan) (group []LogicalPlan, eqEdges []*expression.ScalarFunction, otherConds []expression.Expression) {
-	join, isJoin := p.(*LogicalJoin)
-	if !isJoin || join.preferJoinType > uint(0) || join.JoinType != InnerJoin || join.StraightJoin {
-		return []LogicalPlan{p}, nil, nil
-	}
-
-	lhsGroup, lhsEqualConds, lhsOtherConds := extractJoinGroup(join.children[0])
-	rhsGroup, rhsEqualConds, rhsOtherConds := extractJoinGroup(join.children[1])
-
-	group = append(group, lhsGroup...)
-	group = append(group, rhsGroup...)
-	eqEdges = append(eqEdges, join.EqualConditions...)
-	eqEdges = append(eqEdges, lhsEqualConds...)
-	eqEdges = append(eqEdges, rhsEqualConds...)
-	otherConds = append(otherConds, join.OtherConditions...)
-	otherConds = append(otherConds, lhsOtherConds...)
-	otherConds = append(otherConds, rhsOtherConds...)
-	return group, eqEdges, otherConds
-}
-
-type joinReOrderGreedySolver struct {
-}
-
 type joinReorderGreedySingleGroupSolver struct {
-	ctx          sessionctx.Context
-	curJoinGroup []*jrNode
-	eqEdges      []*expression.ScalarFunction
-	otherConds   []expression.Expression
-}
-
-type jrNode struct {
-	p       LogicalPlan
-	cumCost float64
+	*baseSingleGroupJoinOrderSolver
+	eqEdges []*expression.ScalarFunction
 }
 
 // solve reorders the join nodes in the group based on a greedy algorithm.
@@ -104,14 +67,6 @@ func (s *joinReorderGreedySingleGroupSolver) solve(joinNodePlans []LogicalPlan) 
 	return s.makeBushyJoin(cartesianGroup), nil
 }
 
-func (s *joinReorderGreedySingleGroupSolver) baseNodeCumCost(groupNode LogicalPlan) float64 {
-	cost := groupNode.statsInfo().RowCount
-	for _, child := range groupNode.Children() {
-		cost += s.baseNodeCumCost(child)
-	}
-	return cost
-}
-
 func (s *joinReorderGreedySingleGroupSolver) constructConnectedJoinTree() (*jrNode, error) {
 	curJoinTree := s.curJoinGroup[0]
 	s.curJoinGroup = s.curJoinGroup[1:]
@@ -129,7 +84,7 @@ func (s *joinReorderGreedySingleGroupSolver) constructConnectedJoinTree() (*jrNo
 			if err != nil {
 				return nil, err
 			}
-			curCost := curJoinTree.cumCost + newJoin.statsInfo().RowCount + node.cumCost
+			curCost := s.calcJoinCumCost(newJoin, curJoinTree, node)
 			if bestCost > curCost {
 				bestCost = curCost
 				bestJoin = newJoin
@@ -168,89 +123,5 @@ func (s *joinReorderGreedySingleGroupSolver) checkConnectionAndMakeJoin(leftNode
 	if len(usedEdges) == 0 {
 		return nil, nil
 	}
-	newJoin := s.newCartesianJoin(leftNode, rightNode)
-	newJoin.EqualConditions = usedEdges
-	for _, eqCond := range newJoin.EqualConditions {
-		newJoin.LeftJoinKeys = append(newJoin.LeftJoinKeys, eqCond.GetArgs()[0].(*expression.Column))
-		newJoin.RightJoinKeys = append(newJoin.RightJoinKeys, eqCond.GetArgs()[1].(*expression.Column))
-	}
-	for i := len(remainOtherConds) - 1; i >= 0; i-- {
-		cols := expression.ExtractColumns(remainOtherConds[i])
-		if newJoin.schema.ColumnsIndices(cols) != nil {
-			newJoin.OtherConditions = append(newJoin.OtherConditions, remainOtherConds[i])
-			remainOtherConds = append(remainOtherConds[:i], remainOtherConds[i+1:]...)
-		}
-	}
-	return newJoin, remainOtherConds
-}
-
-func (s *joinReorderGreedySingleGroupSolver) makeBushyJoin(cartesianJoinGroup []LogicalPlan) LogicalPlan {
-	resultJoinGroup := make([]LogicalPlan, 0, (len(cartesianJoinGroup)+1)/2)
-	for len(cartesianJoinGroup) > 1 {
-		resultJoinGroup = resultJoinGroup[:0]
-		for i := 0; i < len(cartesianJoinGroup); i += 2 {
-			if i+1 == len(cartesianJoinGroup) {
-				resultJoinGroup = append(resultJoinGroup, cartesianJoinGroup[i])
-				break
-			}
-			newJoin := s.newCartesianJoin(cartesianJoinGroup[i], cartesianJoinGroup[i+1])
-			for i := len(s.otherConds) - 1; i >= 0; i-- {
-				cols := expression.ExtractColumns(s.otherConds[i])
-				if newJoin.schema.ColumnsIndices(cols) != nil {
-					newJoin.OtherConditions = append(newJoin.OtherConditions, s.otherConds[i])
-					s.otherConds = append(s.otherConds[:i], s.otherConds[i+1:]...)
-				}
-			}
-			resultJoinGroup = append(resultJoinGroup, newJoin)
-		}
-		cartesianJoinGroup, resultJoinGroup = resultJoinGroup, cartesianJoinGroup
-	}
-	return cartesianJoinGroup[0]
-}
-
-func (s *joinReorderGreedySingleGroupSolver) newCartesianJoin(lChild, rChild LogicalPlan) *LogicalJoin {
-	join := LogicalJoin{
-		JoinType:  InnerJoin,
-		reordered: true,
-	}.Init(s.ctx)
-	join.SetSchema(expression.MergeSchema(lChild.Schema(), rChild.Schema()))
-	join.SetChildren(lChild, rChild)
-	return join
-}
-
-func (s *joinReOrderGreedySolver) optimize(p LogicalPlan) (LogicalPlan, error) {
-	return s.optimizeRecursive(p.context(), p)
-}
-
-func (s *joinReOrderGreedySolver) optimizeRecursive(ctx sessionctx.Context, p LogicalPlan) (LogicalPlan, error) {
-	var err error
-	curJoinGroup, eqEdges, otherConds := extractJoinGroup(p)
-	if len(curJoinGroup) > 1 {
-		for i := range curJoinGroup {
-			curJoinGroup[i], err = s.optimizeRecursive(ctx, curJoinGroup[i])
-			if err != nil {
-				return nil, err
-			}
-		}
-		groupSolver := &joinReorderGreedySingleGroupSolver{
-			ctx:        ctx,
-			eqEdges:    eqEdges,
-			otherConds: otherConds,
-		}
-		p, err = groupSolver.solve(curJoinGroup)
-		if err != nil {
-			return nil, err
-		}
-		return p, nil
-	}
-	newChildren := make([]LogicalPlan, 0, len(p.Children()))
-	for _, child := range p.Children() {
-		newChild, err := s.optimizeRecursive(ctx, child)
-		if err != nil {
-			return nil, err
-		}
-		newChildren = append(newChildren, newChild)
-	}
-	p.SetChildren(newChildren...)
-	return p, nil
+	return s.newJoinWithEdges(usedEdges, remainOtherConds, leftNode, rightNode)
 }
